--- app/server/module/accounting/router/router.ts
@@ -1,7 +1,0 @@
-import express from 'express';
-
-const router = express.Router();
-
-router.post('/register/admin');
-
-export default router;
